% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HSVSM.example}
\alias{HSVSM.example}
\title{Example data for HSVSM}
\format{
An object of class \code{matrix} with 20 rows and 4 columns.
}
\usage{
HSVSM.example
}
\arguments{
\item{HSVSM.example}{A p*k numeric matrix for the summary statistics of p SNPs and k traits.}
}
\description{
A toy example of the summary statistics for p SNPs and k traits.
}
\examples{
data(HSVSM.example)
}
\keyword{datasets}
