% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVSM.R
\name{HSVSM}
\alias{HSVSM}
\title{HSVS-M for Multi-SNP Multi-Trait Association Test}
\usage{
HSVSM(
  Z,
  cormat = NULL,
  burnin = 1000,
  N = 2000,
  priors = list(NULL),
  initials = list(NULL)
)
}
\arguments{
\item{Z}{A p*k numeric matrix for the summary statistics of p SNPs and k traits.}

\item{cormat}{A k*k numeric matrix for the correlation matrix for k traits. When NULL, an empirical correlation matrix will be calculated using Z, i.e., cormat=cor(Z).}

\item{burnin}{A non-negative integer for the number of burnin MCMC iterations. Default is 1000.}

\item{N}{A positive integer for the number of non-burnin MCMC iterations. Default is 2000.}

\item{priors}{A list of values for hyperparameters. When NULL, default values will be used.}

\item{initials}{A list of values for the initial values for MCMC sampling. When NULL, default values will be used.}
}
\value{
A list that contains:
\describe{
  \item{p}{HSVS-M's Bayesian p-value.}
  \item{weight}{A data frame for HSVS-M's weight estimates for each trait.}
}
}
\description{
Test for association between a set of SNPs and multiple traits using summary statistics.
}
\examples{
data(HSVSM.example)
out<-HSVSM(HSVSM.example)
HSVSM.p<-out$p
HSVSM.weight<-out$weight
}
